requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
altair==4.1.0
appnope==0.1.2
argcomplete==2.0.0; python_version < '3.8'
argon2-cffi-bindings==21.2.0
argon2-cffi==21.3.0
astor==0.8.1
attrs==21.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1; python_version >= '3.6' and python_version < '3.9'
base58==2.1.1
black==21.12b0
bleach==4.1.0
blinker==1.4
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==7.1.2
cycler==0.11.0; python_version >= '3.6'
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
entrypoints==0.3
flake8==4.0.1
fonttools==4.28.5; python_version >= '3.7'
gitdb==4.0.9
gitpython==3.1.24
google-api-core[grpc]==2.3.0
google-auth==2.3.3
google-cloud-aiplatform==1.9.0
google-cloud-bigquery==2.31.0; python_version < '3.11' and python_version >= '3.6'
google-cloud-core==2.2.1
google-cloud-storage==1.43.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
grpcio-status==1.43.0
grpcio==1.43.0
idna==3.3
importlib-metadata==4.2.0; python_version == '3.7' and python_version < '3.8'
importlib-resources==5.4.0; python_version < '3.9'
ipykernel==6.6.0
ipython-genutils==0.2.0
ipython==7.30.1
ipywidgets==7.6.5
japanize-matplotlib==1.1.3
jedi==0.18.1
jinja2==3.0.3
jsonschema==4.2.1
jupyter-client==7.1.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2; python_version >= '3.7'
markupsafe==2.0.1
matplotlib-inline==0.1.3
matplotlib==3.5.1
mccabe==0.6.1
mistune==0.8.4
mypy-extensions==0.4.3
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.6
numpy==1.21.4
packaging==21.3; python_version >= '3.6'
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
pillow==8.4.0
platformdirs==2.4.0
prometheus-client==0.12.0
prompt-toolkit==3.0.24
proto-plus==1.19.8; python_version >= '3.6'
protobuf==3.19.1
ptyprocess==0.7.0
pyarrow==6.0.1
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycodestyle==2.8.0
pycparser==2.21
pydeck==0.7.1
pyflakes==2.4.0
pygments==2.10.0
pympler==0.9
pyparsing==3.0.6
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz-deprecation-shim==0.1.0.post0
pytz==2021.3
pyzmq==22.3.0
requests==2.26.0
rsa==4.8
send2trash==1.8.0
six==1.16.0
smmap==5.0.0
streamlit==1.2.0
terminado==0.12.1
testpath==0.5.0
toml==0.10.2
tomli==1.2.2
toolz==0.11.2
tornado==6.1
traitlets==5.1.1
typed-ast==1.5.1; python_version < '3.8' and implementation_name == 'cpython'
typing-extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
validators==0.18.2
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.2
zipp==3.7.0; python_version < '3.10'
